+++ split/sub_1007A6A0.c
@@ -1,5 +1,5 @@
 //----- (1007A6A0) --------------------------------------------------------
-int __cdecl sub_1007A6A0(HWND hWnd, LONG dwNewLong)
+int __cdecl sub_1007A6A0(HWND hWnd, intptr_t dwNewLong)
 {
   bool isActiveWindow; // zf
   LONG v3; // eax
